+++ src/CMakeLists.txt
@@ -32,6 +32,7 @@ add_subdirectory(analyticfunctions)
 add_subdirectory(chemistry)
 add_subdirectory(environment)
 add_subdirectory(mrdft)
+add_subdirectory(properties)
 add_subdirectory(qmfunctions)
 add_subdirectory(qmoperators)
 add_subdirectory(scf_solver)
