+++ src/properties/hirshfeld/CMakeLists.txt
@@ -1,2 +1,3 @@
 target_sources(mrchem PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/PolyInterpolator.cpp
     )
