+++ src/properties/hirshfeld/PolyInterpolator.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <vector>
 #include <Eigen/Dense>
 #include <fstream>
@@ -5,28 +6,7 @@
 #include <filesystem>
 #include <iostream>
 
-double polynomialInterpolate5(Eigen::VectorXd &x_in, Eigen::VectorXd &y_in, double x){
-    double xm2 = x_in(0);
-    double xm1 = x_in(1);
-    double x00 = x_in(2);
-    double xp1 = x_in(3);
-    double xp2 = x_in(4);
-
-    double ym2 = y_in(0);
-    double ym1 = y_in(1);
-    double y00 = y_in(2);
-    double yp1 = y_in(3);
-    double yp2 = y_in(4);
-    return ym2 + (x - xm2)*((ym1 - ym2)/(xm1 - xm2) +
-        (x - xm1)*(((y00 - ym1)/(x00 - xm1) + (-ym1 + ym2)/(xm1 - xm2))/(x00 - xm2) +
-        (x - x00)*((-(((y00 - ym1)/(x00 - xm1) + (-ym1 + ym2)/(xm1 - xm2))/(x00 - xm2)) +
-        ((-y00 + ym1)/(x00 - xm1) + (y00 - yp1)/(x00 - xp1))/(-xm1 + xp1))/(-xm2 + xp1) +
-        ((x - xp1)*(-((-(((y00 - ym1)/(x00 - xm1) + (-ym1 + ym2)/(xm1 - xm2))/(x00 - xm2)) +
-        ((-y00 + ym1)/(x00 - xm1) + (y00 - yp1)/(x00 - xp1))/(-xm1 + xp1))/
-        (-xm2 + xp1)) + (-(((-y00 + ym1)/(x00 - xm1) + (y00 - yp1)/(x00 - xp1))/
-        (-xm1 + xp1)) + ((-y00 + yp1)/(x00 - xp1) + (yp1 - yp2)/(xp1 - xp2))/(-x00 + xp2))/
-        (-xm1 + xp2)))/(-xm2 + xp2))));
-}
+double polynomialInterpolate5(Eigen::VectorXd &x_in, Eigen::VectorXd &y_in, double x);
 
 double polynomialInterpolate5_deriv(Eigen::VectorXd &x_in, Eigen::VectorXd &y_in, double x) {
     double xm2 = x_in(0);
@@ -81,12 +61,13 @@ class PolyInterpolator {
         n = x_in.size();
         xmin = x_in(0);
         xmax = x_in(x_in.size() - 1);
-        Eigen::VectorXd x_in(5), y_in(5);
+
+        Eigen::VectorXd x_in_poly(5), y_in_poly(5);
         for (int i = n - 5; i < n; i++) {
-            x_in(i) = x(i);
-            y_in(i) = y(i);
+            x_in_poly(i) = x(i);
+            y_in_poly(i) = y(i);
         }
-        ypxmax = polynomialInterpolate5_deriv(x_in, y_in, xmax);
+        ypxmax = polynomialInterpolate5_deriv(x_in_poly, y_in_poly, xmax);
     }
     
     double evalf(const double &x) const {
